% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sae_mcmc.R
\name{endes_bootstrap_ic}
\alias{endes_bootstrap_ic}
\title{endes_bootstrap_ic}
\usage{
endes_bootstrap_ic(diseno, rep = 500)
}
\arguments{
\item{diseno}{Objeto \code{survey.design2}.}

\item{rep}{Entero. Numero de redisenios bootstrap a simular. Default: 500.}
}
\value{
Un objeto \code{svyrep.design} del tipo bootstrap.
}
\description{
Implementacion computacional intensiva de remuestreo (Bootstrapping) para encuestas
complejas (Rao-Wu-Yue), evadiendo la linearizacion de Taylor para los ICs. Ideal para estadisticos
medianos o percentiles donde la asintotica falla.
}
