% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospatial_api.R
\name{endes_etl_matriz_adyacencia}
\alias{endes_etl_matriz_adyacencia}
\title{endes_etl_matriz_adyacencia}
\usage{
endes_etl_matriz_adyacencia(ruta_shp_distritos, ruta_shp_centros = NULL)
}
\arguments{
\item{ruta_shp_distritos}{Ruta al shapefile de distritos (.shp).}

\item{ruta_shp_centros}{Ruta al shapefile de centros poblados (.shp). Opcional.}
}
\value{
Objeto \code{sf} curado con la topologia correcta, uniendo UBIGEOs para la
inferencia de Modelos Jerarquicos Anidados.
}
\description{
Conecta y procesa los shapefiles de distritos y centros poblados descargados
para usarlos como covariables espaciales (efectos fijos y estructura espacial CAR/SAR)
usando el UBIGEO como llave principal.
}
